Eliminating system crash when changing count products. Added comments for better understanding the code

+++ app/src/main/java/ru/maxstelmakh/ordersfordriver/data/orderApi/model/Goods.kt
@@ -7,6 +7,6 @@ data class Goods(
     @SerializedName("article") var article: Long,
     @SerializedName("name") var name: String,
     @SerializedName("quantity") var quantity: Int,
-    @SerializedName("price") var price: Float,
-    @SerializedName("summ") var summ: Float
+    @SerializedName("price") var price: Double,
+    @SerializedName("summ") var summ: Double
 ) : Serializable
